moveonly: add mp/type-pointer.h
