+++ CMakeLists.txt
@@ -47,6 +47,7 @@ set(MP_PUBLIC_HEADERS
   include/mp/proxy.h
   include/mp/type-context.h
   include/mp/type-optional.h
+  include/mp/type-pointer.h
   include/mp/util.h)
 add_library(multiprocess STATIC
   ${MP_PROXY_SRCS}
